+++ misc/utils.sh
@@ -342,14 +342,14 @@ format_backend_fix() {
 
 lint_backend() {
 	printf "lint backend\\n"
-	cargo clippy --workspace --no-deps --all-targets -- -D warnings || error "clippy failed"
+	cargo clippy --workspace --no-deps --all-targets -- -Dwarnings || error "clippy failed"
 	git ls-files | grep \.sh$ | xargs shellcheck || error "shellcheck failed"
 }
 
 lint_backend_fix() {
 	printf "lint backend\\n"
-	cargo clippy --workspace --no-deps --all-targets --fix --allow-staged --allow-dirty -- -D warnings || error "clippy failed"
-	cargo clippy --workspace --no-deps --all-targets -- -D warnings || error "clippy failed"
+	cargo clippy --workspace --no-deps --all-targets --fix --allow-staged --allow-dirty || error "clippy failed"
+	cargo clippy --workspace --no-deps --all-targets -- -Dwarnings || error "clippy failed"
 	git ls-files | grep \.sh$ | xargs shellcheck || error "shellcheck failed"
 }